move this analysis to its own file
